requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

# keystone dependencies
pbr>=0.6,!=0.7,<1.0
WebOb>=1.2.3,<=1.3.1
eventlet>=0.15.1,<=0.15.2
greenlet>=0.3.2,<=0.4.2
netaddr>=0.7.12,<=0.7.13
PasteDeploy>=1.5.0,<=1.5.2
Paste<=1.7.5.1
Routes>=1.12.3,!=2.0,<=2.1
six>=1.7.0,<=1.9.0
SQLAlchemy>=0.8.4,<=0.9.99,!=0.9.0,!=0.9.1,!=0.9.2,!=0.9.3,!=0.9.4,!=0.9.5,!=0.9.6
sqlalchemy-migrate==0.9.1
passlib<=1.6.2
iso8601>=0.1.9,<=0.1.10
python-keystoneclient>=0.10.0,<=1.1.0
python-slugify==1.0.2
keystonemiddleware>=1.0.0,<1.4.0
oslo.concurrency<=1.2.1         # Apache-2.0
oslo.config>=1.4.0,<=1.6.0 # Apache-2.0
oslo.messaging>=1.4.0,<1.5.0
oslo.db>=1.0.0,<1.1  # Apache-2.0
oslo.i18n>=1.0.0,<=1.3.1 # Apache-2.0
# only for Keystone tests running against keystoneclient master
oslo.serialization>=1.0.0,<=1.2.0 # Apache-2.0
oslo.utils>=1.0.0,<=1.2.1 # Apache-2.0
Babel>=1.3,<=1.3
oauthlib>=0.6,<=0.7.2
dogpile.cache>=0.5.3,<=0.5.6
jsonschema>=2.0.0,<3.0.0
pycadf>=0.6.0,<0.7.0  # Apache-2.0
posix_ipc<=0.9.9
mysql-python
